1+1

+++ springclouddemo2/conversation/src/main/java/com/kjgs/功能/内置功能/执行过滤指定下标方法.java
@@ -24,11 +24,8 @@ public class 执行过滤指定下标方法 extends 功能抽象 {
     @Override
     public void 功能() {
         List 过滤的对象 = (List)获取最近的属性值(所有逻辑对象, this.过滤的对象, List.class);
-        int 过滤的下标 = Integer.parseInt(获取最近的属性值(所有逻辑对象, this.过滤的下标));
+        int 过滤的下标 = (int)Double.parseDouble(获取最近的属性值(所有逻辑对象, this.过滤的下标));
         Object 过滤的结果 = 过滤的对象.get(过滤的下标);
-
-
-
         动作结果=过滤的结果;
     }
 }
